CustomTable responsive styles added
