+++ src/Styles/Categories/CategoriesList/Backoffice_ListCategories.module.css
@@ -1,5 +1,11 @@
 .listContainer {
-  width: 60%;
+  width: 90%;
   max-width: 900px;
   margin: 0 auto;
+}
+
+@media (min-width: 992px) {
+  .listContainer {
+    width: 60%;
+  }
 }
